requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.12.0,<4.0
django-cors-headers>=3.7.0,<4.0
djangorestframework-camel-case>=1.2.0,<2.0
djangorestframework-simplejwt>=5.0.0,<6.0
drf-spectacular>=0.25.0,<1.0
celery>=5.2.0,<6.0
factory_boy>=3.2.0,<4.0
gunicorn>=20.0.0,<21.0
kombu[azure]>=5.2.0,<6.0
psycopg2
azure-servicebus
google-auth>=2.0.0,<3.0
google-auth-oauthlib>=0.4.0,<1.0
cryptography>=3.4.0,<4.0
fcm-django>=2.2.1,<3.0
